requirements.txt
fastapi
tortoise-orm[asyncpg]
uvicorn
aerich
orjson
fastapi-pagination[all]